feat(schedule): add filtered schedule hook
